clang 12 build

+++ base/common/wide_integer_impl.h
@@ -224,7 +224,8 @@ struct integer<Bits, Signed>::_impl
 
     constexpr static void wide_integer_from_bultin(integer<Bits, Signed> & self, double rhs) noexcept
     {
-        if ((rhs > 0 && rhs < std::numeric_limits<uint64_t>::max()) || (rhs < 0 && rhs > std::numeric_limits<int64_t>::min()))
+        if ((rhs > 0 && rhs < static_cast<double>(std::numeric_limits<uint64_t>::max()))
+            || (rhs < 0 && rhs > static_cast<double>(std::numeric_limits<int64_t>::min())))
         {
             self = to_Integral(rhs);
             return;
